+++ internal/node/controller/processes_test.go
@@ -13,8 +13,8 @@ import (
 	"testing"
 
 	"github.com/eagraf/habitat-new/core/state/node"
+	"github.com/eagraf/habitat-new/internal/node/api"
 	"github.com/eagraf/habitat-new/internal/node/api/test_helpers"
-	"github.com/eagraf/habitat-new/internal/node/controller/encrypter"
 	"github.com/eagraf/habitat-new/internal/node/hdb"
 	"github.com/eagraf/habitat-new/internal/process"
 	"github.com/stretchr/testify/assert"
@@ -158,10 +158,7 @@ func TestStartProcessHandler(t *testing.T) {
 	}
 
 	// NewCtrlServer restores the initial state
-	ctrl2, err := NewController2(context.Background(), process.NewProcessManager([]process.Driver{mockDriver}), nil, db, nil,
-		nil,
-		&encrypter.NoopEncrypter{},
-	)
+	ctrl2, err := NewController2(context.Background(), process.NewProcessManager([]process.Driver{mockDriver}), nil, db, nil)
 	require.NoError(t, err)
 
 	s, err := NewCtrlServer(context.Background(), &BaseNodeController{}, "", ctrl2, state)
@@ -173,7 +170,7 @@ func TestStartProcessHandler(t *testing.T) {
 	require.Len(t, procs, 0)
 
 	startProcessHandler := http.HandlerFunc(s.StartProcess)
-	startProcessRoute := newRoute(http.MethodPost, "/node/processes", startProcessHandler)
+	startProcessRoute := api.NewBasicRoute(http.MethodPost, "/node/processes", startProcessHandler)
 	handler := middleware.Middleware(startProcessHandler)
 
 	b, err := json.Marshal(StartProcessRequest{
@@ -239,9 +236,10 @@ func TestStartProcessHandler(t *testing.T) {
 	require.NoError(t, err)
 	require.True(t, running)
 
-	listProcessRoute := newRoute(http.MethodGet, "/node/processes/list", http.HandlerFunc(s.ListProcesses))
+	listFn := http.HandlerFunc(s.ListProcesses)
+	listProcessRoute := api.NewBasicRoute(http.MethodGet, "/node/processes/list", listFn)
 	resp = httptest.NewRecorder()
-	handler = middleware.Middleware(listProcessRoute.fn)
+	handler = middleware.Middleware(listFn)
 	handler.ServeHTTP(
 		resp,
 		httptest.NewRequest(
@@ -259,8 +257,8 @@ func TestStartProcessHandler(t *testing.T) {
 	require.Equal(t, http.StatusOK, resp.Result().StatusCode)
 
 	// Test Stop Process
-	stopRoute := newRoute(http.MethodGet, "/node/processes/stop", http.HandlerFunc(s.StopProcess))
-	handler = middleware.Middleware(stopRoute.fn)
+	stopFn := http.HandlerFunc(s.StopProcess)
+	handler = middleware.Middleware(stopFn)
 
 	// Sad Path
 	mockDriver.returnErr = fmt.Errorf("my error")
@@ -273,7 +271,7 @@ func TestStartProcessHandler(t *testing.T) {
 		resp,
 		httptest.NewRequest(
 			http.MethodGet,
-			stopRoute.Pattern(),
+			"/node/processes/stop",
 			bytes.NewReader(b),
 		),
 	)
@@ -290,7 +288,7 @@ func TestStartProcessHandler(t *testing.T) {
 		resp,
 		httptest.NewRequest(
 			http.MethodGet,
-			stopRoute.Pattern(),
+			"/node/processes/stop",
 			bytes.NewReader(b),
 		),
 	)
@@ -302,7 +300,7 @@ func TestStartProcessHandler(t *testing.T) {
 		resp,
 		httptest.NewRequest(
 			http.MethodPost,
-			stopRoute.Pattern(),
+			"/node/processes/stop",
 			bytes.NewReader([]byte("invalid")),
 		),
 	)
@@ -333,7 +331,7 @@ func TestStartProcessHandler(t *testing.T) {
 		resp,
 		httptest.NewRequest(
 			http.MethodGet,
-			stopRoute.Pattern(),
+			"/node/processes/stop",
 			bytes.NewReader(b),
 		),
 	)
@@ -383,8 +381,6 @@ func TestControllerRestoreProcess(t *testing.T) {
 		jsonState: jsonStateFromNodeState(state),
 	},
 		nil,
-		nil,
-		&encrypter.NoopEncrypter{},
 	)
 	require.NoError(t, err)
