+++ internal/node/controller/routes_test.go
@@ -13,7 +13,6 @@ import (
 	types "github.com/eagraf/habitat-new/core/api"
 	"github.com/eagraf/habitat-new/core/state/node"
 	"github.com/eagraf/habitat-new/internal/node/constants"
-	"github.com/eagraf/habitat-new/internal/node/controller/encrypter"
 	"github.com/eagraf/habitat-new/internal/node/controller/mocks"
 	hdb_mocks "github.com/eagraf/habitat-new/internal/node/hdb/mocks"
 	"github.com/eagraf/habitat-new/internal/package_manager"
@@ -179,8 +178,6 @@ func TestGetNodeState(t *testing.T) {
 			schema:    state.Schema(),
 			jsonState: jsonStateFromNodeState(state),
 		}, nil,
-		nil,
-		&encrypter.NoopEncrypter{},
 	)
 	require.NoError(t, err)
 	ctrlServer, err := NewCtrlServer(
