+++ internal/privy/encrypter.go
@@ -1,4 +1,4 @@
-package encrypter
+package privy
 
 import (
 	"crypto/rand"
